+++ lib/solana_web3.dart
@@ -25,6 +25,7 @@ export 'src/sysvar.dart';
 export 'src/transaction/constants.dart';
 export 'src/transaction/transaction.dart';
 export 'src/validator_info.dart';
+export 'src/web_socket_subscription_manager.dart';
 
 
 /// Solana Library