+++ test/playground_test.dart
@@ -1,0 +1,24 @@
+/// Imports
+/// ------------------------------------------------------------------------------------------------
+
+import 'package:flutter_test/flutter_test.dart';
+import 'package:solana_web3/rpc_models/logs_notification.dart';
+import 'package:solana_web3/rpc_models/program_account.dart';
+import 'package:solana_web3/rpc_models/slot_notification.dart';
+
+
+/// Playground Tests
+/// ------------------------------------------------------------------------------------------------
+
+void main() {
+
+  test('playground', () async {
+    final jsonData = {
+      "parent": 75,
+      "root": 44,
+      "slot": 76
+    };
+    final instance = SlotNotification.fromJson(jsonData);
+    print(instance.toJson());
+  });
+}