+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1+2
+
+* Fix subscription handler issues.
+* Add subscription example.
+
 ## 0.0.1+1
 
 * Update code examples.